requirements.txt
# NL-to-SQL Assistant - Core Dependencies
# Minimal requirements for Phase 1 prototype

# Core ML/NLP
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.13.0

# Database
duckdb>=0.9.0
pandas>=2.0.0
sqlparse>=0.4.4

# UI
gradio>=4.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Metrics and Evaluation
nltk>=3.8.0
rouge-score>=0.1.2
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0